test(extract): align citation timeline regression after rebase
